Separate config and uploader in different modules
