+++ exts/edgeimpulse.dataingestion/edgeimpulse/dataingestion/extension.py
@@ -4,81 +4,9 @@ import omni.ext
 import omni.ui as ui
 from omni.kit.window.file_importer import get_file_importer
 import asyncio
-import requests
-import os
-import json
 
-
-class Config:
-    def __init__(self, config_file="config.json"):
-        self.config_file = config_file
-        self.config_data = self.load_config()
-
-    def load_config(self):
-        if os.path.exists(self.config_file):
-            with open(self.config_file, "r") as file:
-                return json.load(file)
-        return {}
-
-    def save_config(self):
-        with open(self.config_file, "w") as file:
-            json.dump(self.config_data, file)
-
-    def get(self, key, default=None):
-        return self.config_data.get(key, default)
-
-    def set(self, key, value):
-        self.config_data[key] = value
-        self.save_config()
-
-
-async def upload_data(api_key, data_folder, dataset, log_callback, on_upload_complete):
-    dataset_types = ["training", "testing", "anomaly"]
-    if dataset not in dataset_types:
-        log_callback(
-            "Error: Dataset type invalid (must be training, testing, or anomaly)."
-        )
-        return
-
-    url = "https://ingestion.edgeimpulse.com/api/" + dataset + "/files"
-    try:
-        for file in os.listdir(data_folder):
-            file_path = os.path.join(data_folder, file)
-            label = os.path.basename(file_path).split(".")[0]
-            if os.path.isfile(file_path):
-                await asyncio.sleep(1)
-                try:
-                    with open(file_path, "rb") as file_data:
-                        res = requests.post(
-                            url=url,
-                            headers={
-                                "x-label": label,
-                                "x-api-key": api_key,
-                                "x-disallow-duplicates": "1",
-                            },
-                            files={
-                                "data": (
-                                    os.path.basename(file_path),
-                                    file_data,
-                                    "image/png",
-                                )
-                            },
-                        )
-                        if res.status_code == 200:
-                            print(f"Uploaded {file_path}")
-                            log_callback(f"Success: {file_path} uploaded successfully.")
-                        else:
-                            print(f"Error uploading {file_path} {res.text}")
-                            log_callback(
-                                f"Error: {file_path} failed to upload. Status Code {res.status_code}: {res.text}"
-                            )
-                except Exception as e:
-                    log_callback(
-                        f"Error: Failed to process {file_path}. Exception: {str(e)}"
-                    )
-    except FileNotFoundError:
-        log_callback("Error: Data Path invalid.")
-    on_upload_complete()
+from .config import Config
+from .uploader import upload_data
 
 
 class EdgeImpulseExtension(omni.ext.IExt):
